% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R, R/utils.R
\name{bundle_delayed}
\alias{bundle_delayed}
\alias{bundle_args}
\title{Bundle Delayed Objects}
\usage{
bundle_delayed(delayed_list)

bundle_args(...)
}
\arguments{
\item{delayed_list}{A list of \code{Delayed} objects to bundle into a single
\code{Delayed} object}

\item{...}{Ignore (this is a convenience function)}
}
\description{
Bundling \code{Delayed} objects builds a single \code{Delayed}
object out of an arbitrary number of input \code{Delayed} objects.
}
\examples{
ident_fun <- function(x) {
  Sys.sleep(0.01)
  x
}
delayed_ident <- delayed_fun(ident_fun)
d_list <- lapply(1:10, delayed_ident)
d_bundle <- bundle_delayed(d_list)
d_bundle$compute(progress = FALSE)
}
\keyword{internal}
