% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\name{delayed}
\alias{delayed}
\alias{delayed_fun}
\title{Generates a delayed version of an expression}
\usage{
delayed(expr)

delayed_fun(fun)
}
\arguments{
\item{expr}{expression to delay}

\item{fun}{function to delay}
}
\description{
Generates a delayed version of an expression
}
