% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_helpers.R
\name{delayed_traceback}
\alias{delayed_traceback}
\title{Get traceback for last delayed error}
\usage{
delayed_traceback(call_stack = NULL, cat = TRUE)
}
\arguments{
\item{call_stack}{a call stack argument (defaults to last delayed error)}

\item{cat}{if TRUE, uses cat to print the traceback}
}
\description{
Get traceback for last delayed error
}
