% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\docType{class}
\name{Delayed}
\alias{Delayed}
\title{Delayed class that manages dependencies and computes when necessary}
\description{
Delayed class that manages dependencies and computes when necessary
}
\examples{
d <- delayed(3 + 4)
methods::is(d, "Delayed")
d$compute()
}
