% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_delayed_error}
\alias{find_delayed_error}
\title{Find error in delayed chain}
\usage{
find_delayed_error(delayed_object)
}
\arguments{
\item{delayed_object}{the object in which an error occured}
}
\description{
Searches through a network of delayed objects for the first object with state "error"
}
