% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{FutureJob}
\alias{FutureJob}
\title{Future Delayed Jobs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FutureJob
}
\description{
A \code{Job} that leverages the \code{future} framework to
evaluate asynchronously.
}
\keyword{datasets}
