% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{Job}
\alias{Job}
\title{A Job encapsulates the act of evaluating a given delayed object in a particular way
\code{SequentialJob}s evaluate immediately, blocking until they are complete
\code{FutureJob}s leverage the \code{future} package to evaluate according to the specified \code{plan}}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Job
}
\description{
A Job encapsulates the act of evaluating a given delayed object in a particular way
\code{SequentialJob}s evaluate immediately, blocking until they are complete
\code{FutureJob}s leverage the \code{future} package to evaluate according to the specified \code{plan}
}
\keyword{datasets}
