% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_delayed.R
\docType{class}
\name{Lrnr_delayed}
\alias{Lrnr_delayed}
\title{Fit/Predict a learner delayed
A wrapper around any learner that delayes computation}
\format{\code{\link{R6Class}} object.}
\usage{
Lrnr_delayed
}
\value{
\code{\link{Lrnr_base}} object with methods for training and prediction
}
\description{
Fit/Predict a learner delayed
A wrapper around any learner that delayes computation
}
\section{Fields}{

\describe{
\item{\code{params}}{A learner to be wrapped}
}}

\section{Methods}{

\describe{
  \item{\code{new(learner)}}{Wraps the \code{learner} in a delayed computation}
  }
}

\keyword{data}
