% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{plot.Delayed}
\alias{plot.Delayed}
\title{Plot Method for Delayed Objects}
\usage{
\method{plot}{Delayed}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Delayed} for which a task dependency graph
will be generated.}

\item{...}{Additional arugments (currently ignored).}
}
\description{
Plot Method for Delayed Objects
}
