% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_shiny.R
\name{plot_delayed_shiny}
\alias{plot_delayed_shiny}
\title{Animated Representation a Task Dependency Structure}
\usage{
plot_delayed_shiny(scheduler)
}
\arguments{
\item{scheduler}{the scheduler to animate}
}
\description{
uses shiny
}
