% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scheduler.R
\docType{class}
\name{Scheduler}
\alias{Scheduler}
\title{Scheduler class that orders compute tasks and dispatches tasks to workers}
\description{
Scheduler class that orders compute tasks and dispatches tasks to workers
}
\examples{
d <- delayed(3 + 4)
sched <- Scheduler$new(d, SequentialJob)
sched$compute()
}
