% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{SequentialJob}
\alias{SequentialJob}
\title{Sequential Delayed Jobs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequentialJob
}
\description{
A \code{Job} that will evaluate immediately (i.e., in a
sequential fashion), blocking the current process until it completes.
}
\keyword{datasets}
