% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R, R/utils.R
\name{bundle_delayed}
\alias{bundle_delayed}
\alias{bundle_args}
\title{Bundle Delayed Objects}
\usage{
bundle_delayed(delayed_list)

bundle_args(...)
}
\arguments{
\item{delayed_list}{A list of \code{Delayed} objects to bundle into a single
\code{Delayed} object}

\item{...}{Ignore (this is a convenience function)}
}
\description{
Bundling \code{Delayed} objects builds a single \code{Delayed}
object out of an arbitrary number of input \code{Delayed} objects.
}
