% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\name{bundle_args}
\alias{bundle_args}
\alias{bundle_delayed}
\title{Bundle delayed objects into single delayed object}
\usage{
bundle_args(...)

bundle_delayed(delayed_list)
}
\arguments{
\item{...}{ignore}

\item{delayed_list}{A list of delayed objects to bundle into a single delayed object}
}
\description{
Bundle delayed objects into single delayed object
}
