% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\name{delayed}
\alias{delayed}
\alias{delayed_fun}
\title{Generates Delayed Version of an Expression}
\usage{
delayed(expr, sequential = FALSE)

delayed_fun(fun, sequential = FALSE)
}
\arguments{
\item{expr}{expression to delay}

\item{sequential}{if TRUE, never parallelize this task}

\item{fun}{function to delay}
}
\description{
Generates Delayed Version of an Expression

A Delayed version of a function may be called to generate
Delayed objects
}
