% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\name{delayed}
\alias{delayed}
\alias{delayed_fun}
\title{Generates a Delayed version of an expression}
\usage{
delayed(expr, sequential = FALSE)

delayed_fun(fun, sequential = FALSE)
}
\arguments{
\item{expr}{expression to delay}

\item{sequential}{if TRUE, never parallelize this task}

\item{fun}{function to delay}
}
\description{
Generates a Delayed version of an expression

Generates a delayed version of a function, that can be then called to generate Delayed objects
}
