% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delayed.R
\docType{class}
\name{Delayed}
\alias{Delayed}
\title{Delayed class that manages dependencies and computes when necessary}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Delayed
}
\description{
Delayed class that manages dependencies and computes when necessary
}
\keyword{datasets}
