% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{FutureJob}
\alias{FutureJob}
\title{Future Delayed Jobs}
\description{
A \code{Job} that leverages the \code{future} framework to
evaluate asynchronously.
}
\examples{
library(future)
plan(multicore, workers = 1)
d <- delayed(3 + 4)
sched <- Scheduler$new(d, FutureJob, nworkers = 1)
}
