% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{Job}
\alias{Job}
\title{Evaluation of Delayed Objects}
\description{
A \code{Job} encapsulates the act of evaluating a given
\code{delayed} object in a particular way. \code{SequentialJob}s evaluate
immediately, blocking the current process until they are complete.
\code{FutureJob}s leverages \pkg{future} to evaluate according to the
specified \code{\link[future]{plan}}.
}
\keyword{internal}
