% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{make_graph}
\alias{make_graph}
\title{Graphical Representation of a Task Dependency Structure}
\usage{
make_graph(delayed_object, graph = NULL, level = 1)
}
\arguments{
\item{delayed_object}{the Delayed object to graph}

\item{graph}{the current graph, usually NULL}

\item{level}{the level of the node to be graphed, usually NULL}
}
\description{
Graphical Representation of a Task Dependency Structure
}
