% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_shiny.R
\name{plot_delayed_shiny}
\alias{plot_delayed_shiny}
\title{Animated Representation a Task Dependency Structure}
\usage{
plot_delayed_shiny(scheduler)
}
\arguments{
\item{scheduler}{the scheduler to animate}
}
\description{
uses shiny
}
\examples{
\donttest{
adder <- function(x, y) {
  x + y
}
delayed_adder <- delayed_fun(adder)
z <- delayed_adder(3, 4)
z2 <- delayed_adder(z, 4)
z2$sequential <- TRUE
z3 <- delayed_adder(z2, z)
plot_delayed_shiny(z3)
}

}
