% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job.R
\docType{class}
\name{SequentialJob}
\alias{SequentialJob}
\title{A Job that will evaluate immediately}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequentialJob
}
\description{
A Job that will evaluate immediately
}
\keyword{datasets}
